+++ statistics/transformFunctions.go
@@ -1,6 +1,6 @@
 package statistics
 
-func ComputeZScore(x float32, mu float32, sigma float32) float32 {
+func ComputeZScore(x float64, mu float64, sigma float64) float64 {
 	if sigma == 0 {
 		return 0
 	}
@@ -8,6 +8,6 @@ func ComputeZScore(x float32, mu float32, sigma float32) float32 {
 	return (x - mu) / sigma
 }
 
-func ComputeMinMaxScore(x float32, min float32, max float32) float32 {
+func ComputeMinMaxScore(x float64, min float64, max float64) float64 {
 	return (x - min) / (max - min)
 }
